Refactor code, to make default settings possible in a similar fashion all modules
