+++ src/app.module.ts
@@ -14,6 +14,7 @@ import {ProtocolService} from "./protocol/protocol.service";
 import {ProtocolController} from "./protocol/protocol.controller";
 import {LoggerMiddleware} from "./utils/logger.middleware";
 import { SettingsModule } from './settings/settings.module';
+import { HelperModule } from './helper/helper.module';
 
 @Module({
   imports: [
@@ -24,6 +25,7 @@ import { SettingsModule } from './settings/settings.module';
     ProtocolModule,
     FlowsModule,
     SettingsModule,
+    HelperModule,
   ],
   providers: [AppService, ProtocolService],
   controllers: [AppController, ProtocolController],
