+++ src/discovery/discovery.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { DiscoveryController } from './discovery.controller';
 import { DiscoveryService } from './discovery.service';
 import { SettingsModule } from '../settings/settings.module';
+import { HelperModule } from '../helper/helper.module';
 
 @Module({
-  imports: [SettingsModule],
+  imports: [SettingsModule, HelperModule],
   controllers: [DiscoveryController],
   providers: [DiscoveryService],
   exports: [DiscoveryService],
