+++ src/helper/helper.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { HelperService } from './helper.service';
+import { SettingsModule } from '../settings/settings.module';
+
+@Module({
+    imports: [ SettingsModule ],
+    providers: [ HelperService ],
+    exports: [ HelperService ]
+})
+export class HelperModule {}
