+++ src/token/token.module.ts
@@ -7,9 +7,10 @@ import { TokenService } from './token.service';
 import TokenController from './token.controller';
 import { DiscoveryModule } from '../discovery/discovery.module';
 import { SettingsModule } from '../settings/settings.module';
+import { HelperModule } from '../helper/helper.module';
 
 @Module({
-  imports: [ConfigModule.forRoot(), DiscoveryModule, SettingsModule],
+  imports: [ConfigModule.forRoot(), DiscoveryModule, SettingsModule, HelperModule],
   controllers: [TokenController],
   providers: [TokenService],
   exports: [TokenService],
